visClusteringByHubsize {visNetwork}	R Documentation

Network visualization clustering options - by hubsize

Description

Network visualization clustering options - by hubsize

Usage

visClusteringByHubsize(graph, size = NULL)

Arguments

graph
: a visNetwork object

size
: Integer. This method checks all nodes in the network and those with a equal or higher amount of edges than specified with size argument. If size is null (defaut), the size will be determined as the average value plus two standard deviations.

Examples

set.seed(124)
nodes <- data.frame(id = 1:10, color = c(rep("blue", 6), rep("red", 3), rep("green", 1)))
edges <- data.frame(from = round(runif(6)*10), to = round(runif(6)*10))

visNetwork(nodes, edges) %>%
 visClusteringByHubsize()
 
visNetwork(nodes, edges) %>%
 visClusteringByHubsize(size = 2)
 

[Package visNetwork version 2.0.9 Index]