visConfigure {visNetwork}	R Documentation

Network visualization configure options

Description

Network visualization configure options. For full documentation, have a look at visDocumentation.

Usage

visConfigure(
  graph,
  enabled = NULL,
  filter = NULL,
  container = NULL,
  showButton = NULL
)

Arguments

graph
: a visNetwork object

enabled
: Boolean. Default to true. Toggle the configuration interface on or off. This is an optional parameter. If left undefined and any of the other properties of this object are defined, this will be set to true.

filter
: String, Array, Boolean, Function. Default to true. When a boolean, true gives you all options, false will not show any. If a string is supplied, any combination of the following is allowed: nodes, edges, layout, interaction, manipulation, physics, selection, renderer. Feel free to come up with a fun seperating character. Finally, when supplied an array of strings, any of the previously mentioned fields are accepted.

container
: DOM element. This allows you to put the configure list in another HTML container than below the network.

showButton
: Boolean. Default to true. Show the generate options button at the bottom of the configurator.

References

See online documentation http://datastorm-open.github.io/visNetwork/

See Also

visConfigure, visTree, visNetworkEditor

Examples

## Not run: 

nodes <- data.frame(id = 1:3, title = paste0("<p>", 1:3,"<br> tooltip</p>"))
edges <- data.frame(from = c(1,2), to = c(1,3))

visNetwork(nodes, edges) 
 visConfigure(enabled = TRUE, filter = "interaction")
 
# using visNetworkEditor
custom_network <- visNetworkEditor(object = network)
custom_network

custom_network <- visNetworkEditor(object = network, filter = "nodes,edges")
custom_network

## End(Not run)

[Package visNetwork version 2.0.9 Index]