visUpdateEdges {visNetwork}	R Documentation

Function to update the information of edges, with shiny only.

Description

Function to update the information of edges, with shiny only. You can also use this function passing new edges. The link is based on id.

Usage

visUpdateEdges(graph, edges, legend = FALSE)

Arguments

graph
: a visNetworkProxy object

edges
: data.frame with the information of edges. See visEdges

"id" : edge id, for update

"from" : node id, begin of the edge

"to" : node id, end of the edge

"label" : label

"value" : size

"title" : tooltip

...

legend
: Boolean. Update edges on legend ? Default to FALSE

References

See online documentation http://datastorm-open.github.io/visNetwork/

See Also

visNodes for nodes options, visEdges for edges options, visGroups for groups options, visLegend for adding legend, visOptions for custom option, visLayout & visHierarchicalLayout for layout, visPhysics for control physics, visInteraction for interaction, visNetworkProxy & visFocus & visFit for animation within shiny, visDocumentation, visEvents, visConfigure ...

Examples

## Not run: 

# have a look to : 
shiny::runApp(system.file("shiny", package = "visNetwork"))

## End(Not run)

[Package visNetwork version 2.0.3 Index]