visIgraphLayout {visNetwork}	R Documentation

Use a igraph layout for compute coordinates & fast rendering

Description

Use a igraph layout for compute coordinates and fast rendering. This function affect x and y coordinates to nodes data.frame using a igraph layout, and then render network faster with no stabilization. We set some options as : visNodes(physics = FALSE) & visEdges(smooth = FALSE) & visPhysics(stabilization= FALSE), but you can overwrite them using arguments or by add another call after visIgraphLayout

Usage

visIgraphLayout(
  graph,
  layout = "layout_nicely",
  physics = FALSE,
  smooth = FALSE,
  type = "square",
  randomSeed = NULL,
  layoutMatrix = NULL,
  ...
)

Arguments

graph
: a visNetwork object

layout
: Character Name of igraph layout function to use. Default to "layout_nicely"

physics
: Boolean. Default to FALSE. Enabled physics on nodes ?

smooth
: Boolean. Default to FALSE. Use smooth edges ?

type
: Character Type of scale from igrah to vis.js. "square" (default) render in a square limit by height. "full" use width and height to scale in a rectangle.

randomSeed
: Number. The nodes are randomly positioned initially. This means that the settled result is different every time. If you provide a random seed manually, the layout will be the same every time.

layoutMatrix
: in case of layout = 'layout.norm'. the 'layout' argument (A matrix with two or three columns, the layout to normalize)

...
: Adding arguments to layout function

References

See online documentation http://datastorm-open.github.io/visNetwork/

See Also

visNodes for nodes options, visEdges for edges options, visGroups for groups options, visLegend for adding legend, visOptions for custom option, visLayout & visHierarchicalLayout for layout, visPhysics for control physics, visInteraction for interaction, visNetworkProxy & visFocus & visFit for animation within shiny, visDocumentation, visEvents, visConfigure ...

Examples

## Not run: 
##D nnodes <- 200
##D nnedges <- 400
##D 
##D nodes <- data.frame(id = 1:nnodes)
##D edges <- data.frame(from = sample(1:nnodes, nnedges, replace = T), 
##D                    to = sample(1:nnodes, nnedges, replace = T))
##D 
##D # with default layout
##D visNetwork(nodes, edges) %>% 
##D  visIgraphLayout()
##D 
##D # use full space
##D visNetwork(nodes, edges) %>% 
##D  visIgraphLayout(type = "full")
##D 
##D # in circle ?
##D visNetwork(nodes, edges) %>% 
##D  visIgraphLayout(layout = "layout_in_circle") %>%
##D  visOptions(highlightNearest = list(enabled = T, hover = T), 
##D    nodesIdSelection = T)
##D  
##D # keep physics with smooth curves ?
##D visNetwork(nodes, edges) %>% 
##D  visIgraphLayout(physics = TRUE, smooth = TRUE)
##D 
##D # fix randomSeed to keep position
##D visNetwork(nodes, edges) %>% 
##D  visIgraphLayout(randomSeed = 123)
##D  
##D visNetwork(nodes, edges) %>% 
##D  visIgraphLayout(randomSeed = 123)
##D 
##D # layout_with_sugiyama
##D nodes <- data.frame(id = 1:5)
##D edges <- data.frame(from = c(1, 2, 2, 4), to = c(2, 3, 4, 5))
##D 
##D visNetwork(nodes, edges) %>%
##D  visIgraphLayout(layout = "layout_with_sugiyama", layers = c(1, 2, 3, 3, 4))
##D 
##D visNetwork(nodes, edges) %>%
##D  visIgraphLayout(layout = "layout_with_sugiyama")
##D  
## End(Not run)

[Package visNetwork version 2.1.0 Index]