visUpdateNodes {visNetwork}	R Documentation

Function to update the information of nodes, with shiny only.

Description

Function to update the information of nodes, with shiny only. You can also use this function passing new nodes. The link is based on id.

Usage

visUpdateNodes(graph, nodes, updateOptions = TRUE, legend = FALSE)

Arguments

graph
: a visNetworkProxy object

nodes
: data.frame with the information of nodes. Needed at least column "id". See visNodes

"id" : id of the node, needed in the definition of edges and for update nodes

"label" : label of the node

"group" : group of the node. Groups can be configure with visGroups

"value" : size of the node

"title" : tooltip of the node

...

updateOptions
: Boolean. Update options (nodesIdSelection & selectedBy) if needed ? Default to TRUE.

legend
: Boolean. Update nodes on legend ? Default to FALSE

References

See online documentation http://datastorm-open.github.io/visNetwork/

See Also

visNodes for nodes options, visEdges for edges options, visGroups for groups options, visLegend for adding legend, visOptions for custom option, visLayout & visHierarchicalLayout for layout, visPhysics for control physics, visInteraction for interaction, visNetworkProxy & visFocus & visFit for animation within shiny, visDocumentation, visEvents, visConfigure ...

Examples

## Not run: 
##D 
##D # have a look to : 
##D shiny::runApp(system.file("shiny", package = "visNetwork"))
##D 
## End(Not run)

[Package visNetwork version 2.1.0 Index]